requirements.txt
numpy>=1.23
pandas>=1.5
scikit-learn>=1.2
torch>=2.0
transformers>=4.30
sentencepiece>=0.1.99
tqdm>=4.60
jupyter>=1.0
python-dotenv>=1.0
openai>=1.40.0
faiss-cpu>=1.8.0
matplotlib>=3.8
langgraph>=0.2.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
streamlit>=1.35
altair>=5.0
reportlab>=4.0
python-dotenv
